research/models/augmented-llm/prototype0/requirements.txt
fastapi
uvicorn
sentence-transformers
Pillow
faiss-cpu
qdrant-client
psycopg2-binary
langdetect
torch
transformers